PGB03-284 IT4 - INT - UC8 - Integrate Each Subsystem

+++ app/src/main/java/es/unex/giiis/golaso/adapters/JugadorEquipoAdapter.java
@@ -27,13 +27,13 @@ public class JugadorEquipoAdapter extends RecyclerView.Adapter<JugadorEquipoAdap
     Context context;
     List<Jugador> jugadores;
     List<Jugador> jugadoresEquipo;
-    private final int idEquipo;
+    private final long idEquipo;
 
     //boton switch
 
     private ItemClickListener clickListener;
 
-    public JugadorEquipoAdapter(Context context, List<Jugador> responseList, int idEquipo, ItemClickListener clickListener) {
+    public JugadorEquipoAdapter(Context context, List<Jugador> responseList, long idEquipo, ItemClickListener clickListener) {
         this.context = context;
         this.jugadores = responseList;
         this.jugadoresEquipo = new ArrayList<>();
@@ -42,7 +42,7 @@ public class JugadorEquipoAdapter extends RecyclerView.Adapter<JugadorEquipoAdap
     }
 
     //busca los jugadores del equipo y los ordena por nombre
-    private void buscarJugadoresEquipo(int idEquipo) {
+    private void buscarJugadoresEquipo(long idEquipo) {
 
         for (Jugador j:jugadores) {
             if (j.getIdEquipo() == idEquipo) {
