+++ app/src/main/java/es/unex/giiis/golaso/elementos/EquipoDetailFragment.java
@@ -37,12 +37,12 @@ public class EquipoDetailFragment extends Fragment implements FragmentManager.On
     private String mEstadio;
     private String mEntrenador;
     private String mLogo;
-    private int mId;
+    private long mId;
 
     private Equipo mEquipo;
 
     public static EquipoDetailFragment newInstance(String nombre, String ubicacion, String estadio,
-                                                   String entrenador, String logo, int id) {
+                                                   String entrenador, String logo, long id) {
 
         EquipoDetailFragment fragment = new EquipoDetailFragment();
         Bundle args = new Bundle();
@@ -52,7 +52,7 @@ public class EquipoDetailFragment extends Fragment implements FragmentManager.On
         args.putString(ARG_PARAM3, estadio);
         args.putString(ARG_PARAM4, entrenador);
         args.putString(ARG_PARAM5, logo);
-        args.putInt(ARG_PARAM6, id);
+        args.putLong(ARG_PARAM6, id);
 
         fragment.setArguments(args);
 
@@ -70,7 +70,7 @@ public class EquipoDetailFragment extends Fragment implements FragmentManager.On
         args.putString(ARG_PARAM3, equipo.getEstadio());
         args.putString(ARG_PARAM4, equipo.getEntrenador());
         args.putString(ARG_PARAM5, equipo.getLogo());
-        args.putInt(ARG_PARAM6, equipo.getIdEquipo());
+        args.putLong(ARG_PARAM6, equipo.getIdEquipo());
 
         fragment.setArguments(args);
 
@@ -90,7 +90,7 @@ public class EquipoDetailFragment extends Fragment implements FragmentManager.On
             mEstadio = getArguments().getString(ARG_PARAM3);
             mEntrenador = getArguments().getString(ARG_PARAM4);
             mLogo = getArguments().getString(ARG_PARAM5);
-            mId = getArguments().getInt(ARG_PARAM6);
+            mId = getArguments().getLong(ARG_PARAM6);
 
             mEquipo = new Equipo(mId, mNombre, mEntrenador, mEstadio, mUbicacion, mLogo);
